+++ android/app/src/main/java/com/kasette/app/MainActivity.java
@@ -8,6 +8,7 @@ public class MainActivity extends BridgeActivity {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         registerPlugin(ThemeBarsPlugin.class);
+        registerPlugin(SafExportPlugin.class);
         super.onCreate(savedInstanceState);
     }
 }